+++ pkg/cmd/list.go
@@ -10,10 +10,8 @@ import (
 	"path/filepath"
 	"strings"
 	"sync"
-	"time"
 
-	"github.com/cli/go-gh/v2/pkg/api"
-	"github.com/patrickmn/go-cache"
+	"github.com/wayneashleyberry/gh-arc/pkg/client"
 	"golang.org/x/mod/modfile"
 )
 
@@ -63,62 +61,6 @@ func (ap *archivedPrinter) Count() int {
 	defer ap.mu.Unlock()
 
 	return int(ap.count)
-}
-
-// CachedGitHubClient wraps the GitHub API client and transparently caches repo
-// results.
-type CachedGitHubClient struct {
-	client *api.RESTClient
-	cache  *cache.Cache
-}
-
-// RepoResult contains metadata about a GitHub repository, including its
-// archived status and last push date.
-type RepoResult struct {
-	Archived bool   `json:"archived"`
-	PushedAt string `json:"pushed_at"`
-}
-
-// NewCachedGitHubClient creates a new CachedGitHubClient with a default REST
-// client and an in-memory cache. The cache is used to store repository metadata
-// and reduce redundant API calls. Returns an error if the GitHub API client
-// cannot be created.
-func NewCachedGitHubClient() (*CachedGitHubClient, error) {
-	client, err := api.DefaultRESTClient()
-	if err != nil {
-		return nil, fmt.Errorf("failed to create GitHub API client: %w", err)
-	}
-
-	c := cache.New(1*time.Hour, 2*time.Hour)
-
-	return &CachedGitHubClient{client: client, cache: c}, nil
-}
-
-// GetRepoResult returns the archived status and last push date for a GitHub
-// repository. It transparently caches results to avoid redundant API calls. The
-// repo argument should be in the form "owner/repo".
-func (c *CachedGitHubClient) GetRepoResult(repo string) (RepoResult, error) {
-	if cached, found := c.cache.Get(repo); found {
-		return cached.(RepoResult), nil
-	}
-
-	ownerRepo := strings.Split(repo, "/")
-	if len(ownerRepo) != 2 {
-		return RepoResult{}, fmt.Errorf("invalid repo: %s", repo)
-	}
-
-	var result RepoResult
-
-	path := fmt.Sprintf("repos/%s/%s", ownerRepo[0], ownerRepo[1])
-
-	err := c.client.Get(path, &result)
-	if err != nil {
-		return RepoResult{}, fmt.Errorf("failed to fetch repo %s: %w", repo, err)
-	}
-
-	c.cache.Set(repo, result, cache.DefaultExpiration)
-
-	return result, nil
 }
 
 // ListArchivedGoModules lists archived Go modules, optionally including
@@ -203,7 +145,7 @@ func ListArchivedGoModules(ctx context.Context, checkIndirect bool) (int, error)
 		return 0, nil
 	}
 
-	client, err := NewCachedGitHubClient()
+	client, err := client.New()
 	if err != nil {
 		return 0, fmt.Errorf("failed to create github api client: %w", err)
 	}
